reworking optimizer around contract
